Rename graphqlsearch -> savedsearch
Summary:

Test Plan:
